[HVR] Move library implementations to separate packages
Move the implementation of each system functionality to a separate
package, so it is easier to integrate them in other flavours.
