+++ app/src/hvr/java/com/igalia/wolvic/PlatformActivity.java
@@ -32,8 +32,11 @@ import com.huawei.hms.push.RemoteMessage;
 import com.huawei.hvr.LibUpdateClient;
 import com.igalia.wolvic.browser.PermissionDelegate;
 import com.igalia.wolvic.browser.SettingsStore;
+import com.igalia.wolvic.geolocation.HVRLocationManager;
+import com.igalia.wolvic.messaging.WolvicHmsMessageService;
 import com.igalia.wolvic.speech.SpeechRecognizer;
 import com.igalia.wolvic.speech.SpeechServices;
+import com.igalia.wolvic.telemetry.HVRTelemetry;
 import com.igalia.wolvic.telemetry.TelemetryService;
 import com.igalia.wolvic.ui.adapters.SystemNotification;
 import com.igalia.wolvic.ui.widgets.SystemNotificationsManager;
