+++ app/src/hvr/java/com/igalia/wolvic/speech/HVRSpeechRecognizer.java
@@ -1,4 +1,4 @@
-package com.igalia.wolvic;
+package com.igalia.wolvic.speech;
 
 import android.app.Activity;
 import android.content.Context;
