+++ app/src/hvr/java/com/igalia/wolvic/telemetry/HVRTelemetry.java
@@ -1,4 +1,4 @@
-package com.igalia.wolvic;
+package com.igalia.wolvic.telemetry;
 
 import android.content.Context;
 import android.os.Bundle;
